-LongJump from bank to bank. -Up+A on controller 2 performs a hard reset.
Implemented a LongJump function that will use a 6-byte
table entry to allow jumping from bank to bank.
The MMC3 version of SwapPRG was modified to store the
bank being swapped to to $60FC, so LongJump can
automatically swap back to the previous bank.

ReadJoypadData and ProcessJoyButtons were moved to bank 0F
to make room for LongJump and several jump table entries.
Because the controller handling routines were moved, they
were also able to be expanded to allow for a hard reset to be triggered
from the 2nd controller using Up+A.